+++ handy-tarifvergleich-server/Models/Dto/OfferDto.cs
@@ -5,10 +5,10 @@
         public string Name { get; set; }
         public string Provider { get; set; }
         public string OfferURL { get; set; }
-        public double BasePrice { get; set; }
+        public decimal BasePrice { get; set; }
         public bool WorldOffer { get; set; }
         public OfferCostDto Cost { get; set; }
         public OfferDeductionsDto Deductions { get; set; }
-        public int ActivationFee { get; set; }
+        public decimal  ActivationFee { get; set; }
     }
 }
